src/controllers/beaches.ts
import { Controller, Post, ClassMiddleware } from '@overnightjs/core';
import { Request, Response } from 'express';
import Beach from '@src/models/beach';
import { BaseController } from '.';
import AuthMiddleware from '@src/middlewares/auth';

@Controller('beaches')
@ClassMiddleware(AuthMiddleware)
export class BeachesController extends BaseController {
  @Post('')
  public async create(req: Request, res: Response): Promise<void> {
    try {
      const beach = new Beach({ ...req.body, user: req.decoded?.id });
      const result = await beach.save();
      res.status(201).send(result);
    } catch (error) {
      this.sendCreatedUpdatedErrorResponse(res, error);
    }
  }
}
